requirements.txt
Flask
transformers
faiss-cpu
torch
elasticsearch
